server/requirements.txt
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
polygon-api-client==1.13.4
pymongo==4.6.2
python-dotenv==1.0.1
urllib3==1.26.18
websockets==12.0
Werkzeug==3.0.1
